moved navbar, sidebar, footer to widgets
