+++ src/widgets/sidebar/model/const/sidebar-link-props.tsx
