+++ src/widgets/sidebar/model/types/sidebar-link.ts
