+++ src/widgets/sidebar/ui/sidebar-link-group/sidebar-link-group.module.scss
