+++ src/widgets/sidebar/ui/sidebar.module.scss
